+++ pkg/models/all_urls_test.go
@@ -132,53 +132,6 @@ func TestAllUrlsSearchVersionRequirement(t *testing.T) {
 	}
 }
 
-func TestAllUrlsSearchVersionRange(t *testing.T) {
-	defer testutils.SetupTestRulesetsDir(t)()
-
-	err := zlog.NewSugaredDevLogger()
-	if err != nil {
-		t.Fatalf("an error '%s' was not expected when opening a sugared logger", err)
-	}
-	defer zlog.SyncZap()
-	ctx := ctxzap.ToContext(context.Background(), zlog.L)
-	s := ctxzap.Extract(ctx).Sugar()
-	db := sqliteSetup(t) // Setup SQL Lite DB
-	defer CloseDB(db)
-
-	err = LoadTestSQLData(db, ctx)
-	if err != nil {
-		t.Fatalf("failed to load SQL test data: %v", err)
-	}
-	myConfig, err := myconfig.NewServerConfig(nil)
-	if err != nil {
-		t.Fatalf("failed to load Config: %v", err)
-	}
-	myConfig.Database.Trace = true
-	allUrlsModel := NewAllURLModel(db)
-	allUrls, err := allUrlsModel.GetUrlsByPurlNameTypeInRange(ctx, s, "scanoss/engine", "github", ">2.0")
-	if err != nil {
-		t.Errorf("all_urls.GetUrlsByPurlNameTypeInRange() error = %v", err)
-	}
-	if len(allUrls) == 0 {
-		t.Errorf("all_urls.GetUrlsByPurlString() No URLs returned from query")
-	}
-	fmt.Printf("All Urls Version: %#v\n", allUrls)
-
-	_, err = allUrlsModel.GetUrlsByPurlNameTypeInRange(ctx, s, "scanoss/engine", "github", "")
-	if err == nil {
-		t.Errorf("expected error all_urls.GetUrlsByPurlNameTypeInRange() ")
-	}
-
-	_, err = allUrlsModel.GetUrlsByPurlNameTypeInRange(ctx, s, "", "github", ">2.0")
-	if err == nil {
-		t.Errorf("Expected all_urls.GetUrlsByPurlNameTypeInRange() error = %v", err)
-	}
-	_, err = allUrlsModel.GetUrlsByPurlNameTypeInRange(ctx, s, "scanoss/engine", "", ">2.0")
-	if err == nil {
-		t.Errorf("Expected all_urls.GetUrlsByPurlNameTypeInRange() error = %v", err)
-	}
-}
-
 func TestAllUrlsSearchPurlList(t *testing.T) {
 	defer testutils.SetupTestRulesetsDir(t)()
 
@@ -209,37 +162,6 @@ func TestAllUrlsSearchPurlList(t *testing.T) {
 	if len(allUrls) == 0 {
 		t.Errorf("all_urls.GetUrlsByPurlString() No URLs returned from query")
 	}
-}
-
-func TestAllUrlsClosestVersionRequirement(t *testing.T) {
-	defer testutils.SetupTestRulesetsDir(t)()
-
-	err := zlog.NewSugaredDevLogger()
-	if err != nil {
-		t.Fatalf("an error '%s' was not expected when opening a sugared logger", err)
-	}
-	defer zlog.SyncZap()
-	ctx := ctxzap.ToContext(context.Background(), zlog.L)
-	s := ctxzap.Extract(ctx).Sugar()
-	db := sqliteSetup(t) // Setup SQL Lite DB
-	defer CloseDB(db)
-	err = LoadTestSQLData(db, ctx)
-	if err != nil {
-		t.Fatalf("failed to load SQL test data: %v", err)
-	}
-	myConfig, err := myconfig.NewServerConfig(nil)
-	if err != nil {
-		t.Fatalf("failed to load Config: %v", err)
-	}
-	myConfig.Database.Trace = true
-	// allUrlsModel := NewAllURLModel(db)
-	allUrls := []AllURL{{URLHash: "0", Component: "engine", PurlName: "scanoss/engine", SemVer: "v1.0", PurlType: "github"},
-		{URLHash: "1", Component: "engine", PurlName: "scanoss/engine", SemVer: "v1.1", PurlType: "github"},
-		{URLHash: "2", Component: "engine", PurlName: "scanoss/engine", SemVer: "v1.2", PurlType: "github"},
-		{URLHash: "3", Component: "engine", PurlName: "scanoss/engine", SemVer: "v1.3", PurlType: "github"},
-	}
-	_, err = PickClosestUrls(s, allUrls, "scanoss/engine", "github", "v1.3")
-	fmt.Printf("%+v", allUrls)
 }
 
 func TestAllUrlsSearchNoLicense(t *testing.T) {